Extract {Collection,Name}Canonicalizer classes
